tales-of-sre/p1-prometheus/webapp/requirements.txt
flask-restful
prometheus-client
Werkzeug